chore: Add createDrmInfoFromClearKeys and simplify DrmEngine
