+++ lib/media/drm_engine.js
@@ -1215,56 +1215,8 @@ shaka.media.DrmEngine = class {
       return null;
     }
 
-    const StringUtils = shaka.util.StringUtils;
-    const Uint8ArrayUtils = shaka.util.Uint8ArrayUtils;
-    const keys = [];
-    const keyIds = [];
-
-    clearKeys.forEach((key, keyId) => {
-      let kid = keyId;
-      if (kid.length != 22) {
-        kid = Uint8ArrayUtils.toBase64(
-            Uint8ArrayUtils.fromHex(keyId), false);
-      }
-      let k = key;
-      if (k.length != 22) {
-        k = Uint8ArrayUtils.toBase64(
-            Uint8ArrayUtils.fromHex(key), false);
-      }
-      const keyObj = {
-        kty: 'oct',
-        kid: kid,
-        k: k,
-      };
-
-      keys.push(keyObj);
-      keyIds.push(keyObj.kid);
-    });
-
-    const jwkSet = {keys: keys};
-    const license = JSON.stringify(jwkSet);
-
-    // Use the keyids init data since is suggested by EME.
-    // Suggestion: https://bit.ly/2JYcNTu
-    // Format: https://www.w3.org/TR/eme-initdata-keyids/
-    const initDataStr = JSON.stringify({'kids': keyIds});
-    const initData =
-        shaka.util.BufferUtils.toUint8(StringUtils.toUTF8(initDataStr));
-    const initDatas = [{initData: initData, initDataType: 'keyids'}];
-
-    return {
-      keySystem: 'org.w3.clearkey',
-      licenseServerUri: 'data:application/json;base64,' + window.btoa(license),
-      distinctiveIdentifierRequired: false,
-      persistentStateRequired: false,
-      audioRobustness: '',
-      videoRobustness: '',
-      serverCertificate: null,
-      serverCertificateUri: '',
-      sessionType: '',
-      initData: initDatas,
-      keyIds: new Set(keyIds),
-    };
+    const ManifestParserUtils = shaka.util.ManifestParserUtils;
+    return ManifestParserUtils.createDrmInfoFromClearKeys(clearKeys);
   }
 
   /**
